refactor : alimo flow layout
